Move `fj::Transform` to dedicated module
