+++ crates/fj/src/lib.rs
@@ -23,8 +23,9 @@ pub mod syntax;
 mod angle;
 mod shape_2d;
 mod shape_3d;
+mod transform;
 
-pub use self::{angle::*, shape_2d::*, shape_3d::*};
+pub use self::{angle::*, shape_2d::*, shape_3d::*, transform::Transform};
 pub use fj_proc::*;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
