+++ crates/fj/src/shape_3d.rs
@@ -1,4 +1,4 @@
-use crate::{Angle, Shape, Shape2d};
+use crate::{Shape, Shape2d, Transform};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
@@ -51,44 +51,6 @@ impl From<Group> for Shape {
 impl From<Group> for Shape3d {
     fn from(shape: Group) -> Self {
         Self::Group(Box::new(shape))
-    }
-}
-
-/// A transformed 3-dimensional shape
-///
-/// # Limitations
-///
-/// Transformations are currently limited to a rotation, followed by a
-/// translation.
-///
-/// See issue:
-/// <https://github.com/hannobraun/Fornjot/issues/101>
-#[derive(Clone, Debug)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[repr(C)]
-pub struct Transform {
-    /// The shape being transformed
-    pub shape: Shape3d,
-
-    /// The axis of the rotation
-    pub axis: [f64; 3],
-
-    /// The angle of the rotation
-    pub angle: Angle,
-
-    /// The offset of the translation
-    pub offset: [f64; 3],
-}
-
-impl From<Transform> for Shape {
-    fn from(shape: Transform) -> Self {
-        Self::Shape3d(Shape3d::Transform(Box::new(shape)))
-    }
-}
-
-impl From<Transform> for Shape3d {
-    fn from(shape: Transform) -> Self {
-        Self::Transform(Box::new(shape))
     }
 }
